+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Do not initialise ReportList vue component if it is not shown !1159 @dthulke
 - Ensures quiz break message after three failures inbetween 30 days #736 !1162 @svenpascal
 - Prevent forum thread email sending to countries and federal states !1160 @jofranz
+- Prefetchtime is now correctly stored when creating a new store !1170 @dthulke
 
 ## Refactoring
 - Removed support for old passwords stored in sha1 or md5, since we switched to Argon2 now almost 2 years ago. !1095 @alangecker